fix cmake files (CMakeLists.txt) complete Query 1.3 fix many things in the other queries checkAndDecodeAN now returns a tuple with 5 elements: 0=bat, 1,2=vector<bool>*'s for encoded heads / tails (or null otherwise) added macros in ssbm.h to delete the vector<bool>*'s returned by the *AN-operators (with proper nullptr-checks) run.sh now prints to stdout the current state

+++ include/column_storage/ColumnBatIterator.h
@@ -92,7 +92,7 @@ public:
 
     /** @return true if a next item is available - otherwise false */
     virtual bool hasNext() override {
-        return mPosition < Csize;
+        return mPosition < static_cast<ssize_t> (Csize);
     }
 
     virtual oid_t head() override {
